+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.1 (October 14, 2022)
+
+FEATURE:
+* Supports Slack Connections (teams or services)
+
+CHORE:
+* Added outputs for Slack Connections
+* Added outputs for Maintenance Windows
+* Improved examples
 ## 1.1.0 (September 27, 2022)
 
 CHORE: